Initial commit
The byebyebios project provides an x86 boot sector that should
be copied to any disk image that does not intend to support
use of BIOS firmware. It will display a message to the user,
on the first serial port and VGA console, informing them of
the requirement to boot using UEFI firmware.

Signed-off-by: Daniel P. Berrangé <berrange@redhat.com>